+++ stubs/DecodeLabs/Archetype.php
@@ -5,30 +5,49 @@
  */
 namespace DecodeLabs;
 
+use DecodeLabs\Archetype\Handler as Inst;
+use DecodeLabs\Archetype\Normalizer as Ref1;
+use DecodeLabs\Archetype\Resolver as Ref0;
 use DecodeLabs\Veneer\Proxy;
 use DecodeLabs\Veneer\ProxyTrait;
-use DecodeLabs\Archetype\Handler as Inst;
-use DecodeLabs\Archetype\Resolver as Ref0;
-use Generator as Ref1;
+use Generator as Ref2;
 
 class Archetype implements Proxy
 {
     use ProxyTrait;
 
-    const VENEER = 'DecodeLabs\Archetype';
-    const VENEER_TARGET = Inst::class;
+    public const VENEER = 'DecodeLabs\Archetype';
+    public const VENEER_TARGET = Inst::class;
 
     public static Inst $instance;
 
-    public static function register(Ref0 $resolver): void {}
-    public static function unregister(Ref0 $resolver): void {}
-    public static function resolve(string $interface, string $name): string {
+    public static function register(Ref0|Ref1 $item): void
+    {
+    }
+    public static function unregister(Ref0|Ref1 $item): void
+    {
+    }
+    public static function registerCustomNormalizer(string $interface, callable $normalizer, int $priority = 10): Ref1
+    {
+        return static::$instance->registerCustomNormalizer(...func_get_args());
+    }
+    public static function extend(string $interface, string $namespace): void
+    {
+    }
+    public static function resolve(string $interface, string $name): string
+    {
         return static::$instance->resolve(...func_get_args());
     }
-    public static function findFile(string $interface, string $name): string {
+    public static function normalize(string $interface, string $name): string
+    {
+        return static::$instance->normalize(...func_get_args());
+    }
+    public static function findFile(string $interface, string $name): string
+    {
         return static::$instance->findFile(...func_get_args());
     }
-    public static function scanClasses(string $interface): Ref1 {
+    public static function scanClasses(string $interface): Ref2
+    {
         return static::$instance->scanClasses(...func_get_args());
     }
-};
+}